requirements.txt
apache-airflow-providers-postgres==1!2.3.0
psycopg2-binary==2.9.1